Disable min/max macros from windows.h

+++ OVP/D3D9Client/D3D9Client.h
@@ -36,7 +36,7 @@
 #define PP_DEFAULT			0x1
 #define PP_LENSFLARE		0x2
 
-#define MAX_SCENE_LIGHTS	24
+#define MAX_SCENE_LIGHTS	(DWORD)24
 #define MAX_MESH_LIGHTS		8	// Must match the setting in D3D9Client.fx
 
 #ifdef _NVAPI_H
